Converts a binary number to an unsigned int

+++ 0x14-bit_manipulation/0-binary_to_uint.c
@@ -10,7 +10,7 @@
  */
 unsigned int binary_to_uint(const char *b)
 {
-	size_t num = 0, mult = 1;
+	unsigned int num = 0, mult = 1;
 	int len;
 
 	if (b == '\0')
